detailed view's final update

+++ TochkaK/LibrarySingleBookOperation.m
@@ -10,7 +10,7 @@
 #import "LibraryBookCreator.h"
 @interface LibrarySingleBookOperation()
 
-@property (assign, nonatomic) int requestedBookID;
+@property (assign, nonatomic) NSInteger requestedBookID;
 
 - (NSURL *)constructURLForBookWithID:(NSInteger)ID ;
